+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - The enclosureType custom field is now required for alternative enclosure for live items.
 - The `enclosure` custom field for live items should be `liveEnclosure` now.
+- The `alternateEnclosure` custom field for live items should be `liveAlternateEnclosure` now.
 
 ## 0.15.0 - 2023-08-08